change type of point to int

+++ src/java/linkjvm/high/vision/Point2.java
@@ -1,15 +1,15 @@
 package linkjvm.high.vision;
 
 public class Point2 {
-	private long x;
-	private long y;
+	private int x;
+	private int y;
 	
 	/**
 	 * 
 	 * @param x
 	 * @param y
 	 */
-	public Point2(long x, long y) {
+	public Point2(int x, int y) {
 		super();
 		this.x = x;
 		this.y = y;
@@ -18,28 +18,28 @@ public class Point2 {
 	/**
 	 * @return the x
 	 */
-	public synchronized long getX() {
+	public synchronized int getX() {
 		return x;
 	}
 
 	/**
 	 * @param x the x to set
 	 */
-	public synchronized void setX(long x) {
+	public synchronized void setX(int x) {
 		this.x = x;
 	}
 
 	/**
 	 * @return the y
 	 */
-	public synchronized long getY() {
+	public synchronized int getY() {
 		return y;
 	}
 
 	/**
 	 * @param y the y to set
 	 */
-	public synchronized void setY(long y) {
+	public synchronized void setY(int y) {
 		this.y = y;
 	}
